delivery-service/requirements.txt
faust-streaming==0.8.8
python-dateutil==2.8.2
sortedcontainers==2.4.0